ios打包支持

+++ lib/common/sound/recorder.dart
@@ -1,13 +1,12 @@
 import 'dart:async';
 import 'dart:io';
-
 import 'package:record/record.dart';
 
 abstract class IRecorder {
   Future<void> init();
 
   Future<bool> start(String filePath, int durationSeconds,
-      Function(String?)? autoCloseCallback);
+      Function(int)? onTickerCallback, Function(String?)? autoCloseCallback);
 
   Future<String?> stop();
 
@@ -16,96 +15,46 @@ abstract class IRecorder {
   Future<void> dispose();
 }
 
-// class SoundRecorder extends IRecorder {
-//   FlutterSoundRecorder flutterSoundRecorder = FlutterSoundRecorder();
-//   StreamSubscription<RecordingDisposition>? recorderSubscription;
-//   String? filePath;
-//
-//   @override
-//   Future<void> init() async {
-//     await flutterSoundRecorder.openRecorder();
-//     await flutterSoundRecorder.setSubscriptionDuration(Duration(seconds: 1));
-//   }
-//
-//   @override
-//   Future<void> start(String filePath, int durationSeconds,
-//       Function(String?)? autoCloseCallback) async {
-//     if (flutterSoundRecorder.isRecording) {
-//       await flutterSoundRecorder.stopRecorder();
-//       recorderSubscription?.cancel();
-//     }
-//     try {
-//       await flutterSoundRecorder.startRecorder(
-//           codec: Codec.pcm16WAV,
-//           sampleRate: 16000,
-//           toFile: filePath,
-//           bitRate: 128,
-//           numChannels: 1);
-//       this.filePath = filePath;
-//       recorderSubscription = flutterSoundRecorder.onProgress?.listen((a) async {
-//         var duration = a.duration.inSeconds;
-//         if (duration >= durationSeconds) {
-//           await stop();
-//           autoCloseCallback?.call(filePath);
-//         }
-//       });
-//     } catch (e) {
-//     }
-//   }
-//
-//   @override
-//   Future<String?> stop() async {
-//     if (flutterSoundRecorder.isRecording) {
-//       await flutterSoundRecorder.stopRecorder();
-//     }
-//     return filePath;
-//   }
-//
-//   @override
-//   Future<void> dispose() async {
-//     filePath = null;
-//     recorderSubscription?.cancel();
-//     await flutterSoundRecorder.closeRecorder();
-//   }
-//
-//   @override
-//   Future<void> cancel() async {
-//     filePath = null;
-//     recorderSubscription?.cancel();
-//     await flutterSoundRecorder.stopRecorder();
-//     if (filePath != null) {
-//       File(filePath!).deleteSync(recursive: true);
-//     }
-//   }
-// }
-
 class RecordRecorder extends IRecorder {
   final record = AudioRecorder();
   Timer? timer;
   String? filePath;
+  int timeCount = 0;
 
   @override
   Future<void> dispose() async {
     record.dispose();
+    timer?.cancel();
+    timeCount = 0;
   }
 
   @override
   Future<void> init() async {}
 
   @override
-  Future<bool> start(String filePath, int durationSeconds,
+  Future<bool> start(
+      String filePath,
+      int durationSeconds,
+      Function(int)? onTickerCallback,
       Function(String? p1)? autoCloseCallback) async {
-    if(await record.hasPermission()){
+    if (await record.hasPermission()) {
       if (await record.isRecording()) {
         record.cancel();
       }
       this.filePath = filePath;
-      timer?.cancel();
-      timer = Timer(Duration(seconds: durationSeconds), () async {
-        await stop();
-        autoCloseCallback?.call(filePath);
-      });
       try {
+        timeCount = 0;
+        timer?.cancel();
+        ///0 1 2 3 4 5
+        timer = Timer.periodic(Duration(seconds: 1), (a) async {
+          if (timeCount >= durationSeconds) {
+            await stop();
+            autoCloseCallback?.call(filePath);
+          } else {
+            timeCount++;
+            onTickerCallback?.call(timeCount);
+          }
+        });
         await record.start(
             const RecordConfig(
                 encoder: AudioEncoder.wav,
@@ -117,10 +66,12 @@ class RecordRecorder extends IRecorder {
                 echoCancel: true),
             path: filePath);
         return true;
-      }catch(e){
+      } catch (e) {
+        timer?.cancel();
+        timeCount = 0;
         return false;
       }
-    }else{
+    } else {
       return false;
     }
   }
@@ -136,8 +87,10 @@ class RecordRecorder extends IRecorder {
   @override
   Future<void> cancel() async {
     await record.cancel();
-    if (filePath != null) {
+    if (filePath != null && File(filePath!).existsSync()) {
       File(filePath!).deleteSync(recursive: true);
     }
+    timer?.cancel();
+    timeCount = 0;
   }
 }
